move mm256_extractf() to a new "avx" module
